Added front end parts
The remaining part that i have to do is to debug some issues i got on the stylesheet. Then remove some questions that take a lot of time to output and replace them by new ones, add the push notification and review the script and adding the aria-label classes for each of these elements. After these changes i will make another pull request